refactor intersection union
