+++ src/operations/intersection_union.js
@@ -2,6 +2,10 @@
 import { mergeFuzzySetProps } from './helpers';
 import FuzzySet from './../fuzzy-set';
 
-export default function union(...fsList) {
+export function intersection(...fsList) {
+  return new FuzzySet(mergeFuzzySetProps(Math.min, false, ...fsList));
+}
+
+export function union(...fsList) {
   return new FuzzySet(mergeFuzzySetProps(Math.max, false, ...fsList));
 }
